Program that prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -16,7 +16,7 @@ int main(void)
 		sum = num_1 + num_2;
 		num_1 = num_2;
 		num_2 = sum;
-		printf("%ld", sum);
+		printf("%ld", (long)sum);
 		if (iterate < 49)
 			printf(", ");
 	}
